qml host
Отображение хостов, их информации и состояния подключения. Удаление после отключения от сервера

+++ IOTV/src/IOTV_server/IOT_Host/IOT_Host.cpp
@@ -268,7 +268,7 @@ void IOT_Host::slotResendDataRead()
 
 void IOT_Host::slotTimeOut()
 {
-    for (size_t i = 0; i <  readChannelLength(); i++)
+    for (int i = 0; i <  readChannelLength(); i++)
         readData(i);
 }
